requirements.txt
asgiref==3.7.2
astroid==2.15.6
autopep8==2.0.2
certifi==2023.7.22
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
filelock==3.12.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
pipenv==2023.9.8
platformdirs==3.10.0
psycopg2==2.9.7
pycodestyle==2.11.0
pylint==2.17.5
pytz==2023.3.post1
sqlparse==0.4.4
tk==0.1.0
tomlkit==0.12.1
tzdata==2023.3
virtualenv==20.24.5
wrapt==1.15.0
